refactor(status): extract cmd_status into cli/status_cmd.rs
Lifts the existing status logic into vitalog::cli::status_cmd::
execute so it can be exercised from unit tests (matching the
trend_cmd pattern). The function now opens an rw connection and
runs materializer::sync_all before reading, mirroring today_cmd
— which ensures just-logged data shows up in `vitalog status`
the same way it shows up in `vitalog today` (issue #27).

Sync errors are swallowed, same as today_cmd.
